Rename project to sbt-findbugs
