+++ src/main/scala/org/scalasbt/findbugs/FindBugsKeys.scala
@@ -1,5 +1,5 @@
 /*
- * This file is part of findbugs4sbt
+ * This file is part of sbt-findbugs
  *
  * Copyright (c) Joachim Hofer & contributors
  * All rights reserved.
@@ -10,22 +10,22 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package de.johoop.findbugs4sbt
+package org.scalasbt.findbugs
 
 import java.io.File
 
-import de.johoop.findbugs4sbt.Effort._
-import de.johoop.findbugs4sbt.Priority._
-import de.johoop.findbugs4sbt.ReportType._
+import org.scalasbt.findbugs.Effort._
+import org.scalasbt.findbugs.Priority.Priority
+import org.scalasbt.findbugs.ReportType._
 import sbt._
 
 import scala.xml.Node
 
-private[findbugs4sbt] case class PathSettings(reportPath: Option[File], analyzedPath: Seq[File], auxPath: Seq[File])
+private[findbugs] case class PathSettings(reportPath: Option[File], analyzedPath: Seq[File], auxPath: Seq[File])
 
-private[findbugs4sbt] case class FilterSettings(includeFilters: Option[Node], excludeFilters: Option[Node])
+private[findbugs] case class FilterSettings(includeFilters: Option[Node], excludeFilters: Option[Node])
 
-private[findbugs4sbt] case class MiscSettings(
+private[findbugs] case class MiscSettings(
     reportType: Option[ReportType],
     priority: Priority,
     onlyAnalyze: Option[Seq[String]],
