+++ src/sbt-test/findbugs4sbt/simple/project/plugins.sbt
@@ -4,5 +4,5 @@
     throw new RuntimeException("""|The system property 'plugin.version' is not defined.
                                   |Specify this property using the parameter -Dplugin.version=<version>""".stripMargin)
 
-  else addSbtPlugin("de.johoop" % "findbugs4sbt" % pluginVersion)
+  else addSbtPlugin("org.scala-sbt" % "sbt-findbugs" % pluginVersion)
 }
